Moved DiffUtils to their Adapter files
